Split up Go types, ApiError implements error
