+++ api/lib_test.go
@@ -186,7 +186,7 @@ func TestMultipleInstances(t *testing.T) {
 
 	// succeed to execute store1 with fred
 	resp = exec(t, cache, id, "fred", store1, api, 0x103df)
-	require.Nil(t, resp.Err, resp.Err.String())
+	require.Nil(t, resp.Err, "%v", resp.Err)
 	require.Equal(t, 1, len(resp.Ok.Messages))
 	logs := resp.Ok.Log
 	require.Equal(t, 2, len(logs))
@@ -207,7 +207,7 @@ func requireOkResponse(t *testing.T, res []byte, expectedMsgs int) {
 	var resp types.CosmosResponse
 	err := json.Unmarshal(res, &resp)
 	require.NoError(t, err)
-	require.Nil(t, resp.Err, resp.Err.String())
+	require.Nil(t, resp.Err, "%v", resp.Err)
 	require.Equal(t, expectedMsgs, len(resp.Ok.Messages))
 }
 
@@ -268,6 +268,6 @@ func TestQuery(t *testing.T) {
 	var qres types.QueryResponse
 	err = json.Unmarshal(data, &qres)
 	require.NoError(t, err)
-	require.Nil(t, qres.Err, qres.Err.String())
+	require.Nil(t, qres.Err, "%v", qres.Err)
 	require.Equal(t, string(qres.Ok), `{"verifier":"fred"}`)
 }
