+++ types/queries.go
@@ -1,0 +1,8 @@
+package types
+
+//-------- Queries --------
+
+type QueryResponse struct {
+	Ok  []byte    `json:"Ok,omitempty"`
+	Err *ApiError `json:"Err,omitempty"`
+}
